AArch64: LP64 fixes to libcore.
Make changes to native and Java code to support 64 bit or 32 bit native
pointers held in Java longs.

In particular, the MemoryTest class, in Java, used ints to do pointer
calculations. This was inadequate and was crashing on 64bit platforms.
The class was changed to use longs for pointer calculations.

Also, the startElement methods in the ExpatParser and EntityParser classes
was changed to use long rather than int for the attributePointer parameter.
The attributePointer attribute of the ExpatParser class was also changed
from int to long. The JNI use of the ExpatParser class's startElement
method was changed accordingly.

+++ dalvik/src/main/java/dalvik/system/DexFile.java
@@ -34,7 +34,7 @@ import libcore.io.StructStat;
  * read-only by the VM.
  */
 public final class DexFile {
-    private int mCookie;
+    private long mCookie;
     private final String mFileName;
     private final CloseGuard guard = CloseGuard.get();
 
@@ -215,7 +215,7 @@ public final class DexFile {
         return defineClass(name, loader, mCookie, suppressed);
     }
 
-    private static Class defineClass(String name, ClassLoader loader, int cookie,
+    private static Class defineClass(String name, ClassLoader loader, long cookie,
                                      List<Throwable> suppressed) {
         Class result = null;
         try {
@@ -232,7 +232,7 @@ public final class DexFile {
         return result;
     }
 
-    private static native Class defineClassNative(String name, ClassLoader loader, int cookie)
+    private static native Class defineClassNative(String name, ClassLoader loader, long cookie)
         throws ClassNotFoundException, NoClassDefFoundError;
 
     /**
@@ -267,7 +267,7 @@ public final class DexFile {
     }
 
     /* return a String array with class names */
-    native private static String[] getClassNameList(int cookie);
+    native private static String[] getClassNameList(long cookie);
 
     /**
      * Called when the class is finalized. Makes sure the DEX file is closed.
@@ -291,20 +291,20 @@ public final class DexFile {
      * Open a DEX file.  The value returned is a magic VM cookie.  On
      * failure, an IOException is thrown.
      */
-    private static int openDexFile(String sourceName, String outputName,
+    private static long openDexFile(String sourceName, String outputName,
         int flags) throws IOException {
         return openDexFileNative(new File(sourceName).getCanonicalPath(),
                                  (outputName == null) ? null : new File(outputName).getCanonicalPath(),
                                  flags);
     }
 
-    native private static int openDexFileNative(String sourceName, String outputName,
+    native private static long openDexFileNative(String sourceName, String outputName,
         int flags) throws IOException;
 
     /*
      * Close DEX file.
      */
-    native private static void closeDexFile(int cookie);
+    native private static void closeDexFile(long cookie);
 
     /**
      * Returns true if the VM believes that the apk/jar file is out of date
